+++ ice_driver/driver/osmm_interface.h
@@ -92,7 +92,7 @@ void cve_osmm_inf_dma_buf_unmap(os_allocation_handle halloc);
 int cve_osmm_dma_buf_map(os_domain_handle *hdomain,
 		struct cve_dma_handle **dma_handle,
 		u32 dma_domain_array_size,
-		u32 size_bytes,
+		u64 size_bytes,
 		union allocation_address alloc_addr,
 		ice_va_t cve_addr,
 		u32 prot,
